Fix database issues
